+++ src/igor/Core/Bestscenarioscounter.cpp
@@ -108,17 +108,17 @@ void Best_scenarios_counter::count_scenario(
 
         // Get mismatches and add them to the mismatch list
         if (scenario.mismatches[V_gene_seq]) {
-            const vector<int> &v_mismatch_list = *scenario.mismatches[V_gene_seq];
+            const vector<size_t> &v_mismatch_list = *scenario.mismatches[V_gene_seq];
             single_scenario_mismatches_list.insert(single_scenario_mismatches_list.end(), v_mismatch_list.begin(),
                                                    v_mismatch_list.end());
         }
         if (scenario.mismatches[D_gene_seq]) {
-            const vector<int> &d_mismatch_list = *scenario.mismatches[D_gene_seq];
+            const vector<size_t> &d_mismatch_list = *scenario.mismatches[D_gene_seq];
             single_scenario_mismatches_list.insert(single_scenario_mismatches_list.end(), d_mismatch_list.begin(),
                                                    d_mismatch_list.end());
         }
         if (scenario.mismatches[J_gene_seq]) {
-            const vector<int> &j_mismatch_list = *scenario.mismatches[J_gene_seq];
+            const vector<size_t> &j_mismatch_list = *scenario.mismatches[J_gene_seq];
             single_scenario_mismatches_list.insert(single_scenario_mismatches_list.end(), j_mismatch_list.begin(),
                                                    j_mismatch_list.end());
         }
@@ -147,17 +147,17 @@ void Best_scenarios_counter::count_scenario(
 
             // Get mismatches and add them to the mismatch list
             if (scenario.mismatches[V_gene_seq]) {
-                const vector<int> &v_mismatch_list = *scenario.mismatches[V_gene_seq];
+                const vector<size_t> &v_mismatch_list = *scenario.mismatches[V_gene_seq];
                 single_scenario_mismatches_list.insert(single_scenario_mismatches_list.end(), v_mismatch_list.begin(),
                                                        v_mismatch_list.end());
             }
             if (scenario.mismatches[D_gene_seq]) {
-                const vector<int> &d_mismatch_list = *scenario.mismatches[D_gene_seq];
+                const vector<size_t> &d_mismatch_list = *scenario.mismatches[D_gene_seq];
                 single_scenario_mismatches_list.insert(single_scenario_mismatches_list.end(), d_mismatch_list.begin(),
                                                        d_mismatch_list.end());
             }
             if (scenario.mismatches[J_gene_seq]) {
-                const vector<int> &j_mismatch_list = *scenario.mismatches[J_gene_seq];
+                const vector<size_t> &j_mismatch_list = *scenario.mismatches[J_gene_seq];
                 single_scenario_mismatches_list.insert(single_scenario_mismatches_list.end(), j_mismatch_list.begin(),
                                                        j_mismatch_list.end());
             }
@@ -196,17 +196,17 @@ void Best_scenarios_counter::count_scenario(
 
         // Get mismatches and add them to the mismatch list
         if (mismatches_lists.exist(V_gene_seq)) {
-            const vector<int> &v_mismatch_list = *mismatches_lists.at(V_gene_seq);
+            const vector<size_t> &v_mismatch_list = *mismatches_lists.at(V_gene_seq);
             single_scenario_mismatches_list.insert(single_scenario_mismatches_list.end(), v_mismatch_list.begin(),
                                                    v_mismatch_list.end());
         }
         if (mismatches_lists.exist(D_gene_seq)) {
-            const vector<int> &d_mismatch_list = *mismatches_lists[D_gene_seq];
+            const vector<size_t> &d_mismatch_list = *mismatches_lists[D_gene_seq];
             single_scenario_mismatches_list.insert(single_scenario_mismatches_list.end(), d_mismatch_list.begin(),
                                                    d_mismatch_list.end());
         }
         if (mismatches_lists.exist(J_gene_seq)) {
-            const vector<int> &j_mismatch_list = *mismatches_lists.at(J_gene_seq);
+            const vector<size_t> &j_mismatch_list = *mismatches_lists.at(J_gene_seq);
             single_scenario_mismatches_list.insert(single_scenario_mismatches_list.end(), j_mismatch_list.begin(),
                                                    j_mismatch_list.end());
         }
@@ -235,17 +235,17 @@ void Best_scenarios_counter::count_scenario(
 
             // Get mismatches and add them to the mismatch list
             if (mismatches_lists.exist(V_gene_seq)) {
-                const vector<int> &v_mismatch_list = *mismatches_lists.at(V_gene_seq);
+                const vector<size_t> &v_mismatch_list = *mismatches_lists.at(V_gene_seq);
                 single_scenario_mismatches_list.insert(single_scenario_mismatches_list.end(), v_mismatch_list.begin(),
                                                        v_mismatch_list.end());
             }
             if (mismatches_lists.exist(D_gene_seq)) {
-                const vector<int> &d_mismatch_list = *mismatches_lists[D_gene_seq];
+                const vector<size_t> &d_mismatch_list = *mismatches_lists[D_gene_seq];
                 single_scenario_mismatches_list.insert(single_scenario_mismatches_list.end(), d_mismatch_list.begin(),
                                                        d_mismatch_list.end());
             }
             if (mismatches_lists.exist(J_gene_seq)) {
-                const vector<int> &j_mismatch_list = *mismatches_lists.at(J_gene_seq);
+                const vector<size_t> &j_mismatch_list = *mismatches_lists.at(J_gene_seq);
                 single_scenario_mismatches_list.insert(single_scenario_mismatches_list.end(), j_mismatch_list.begin(),
                                                        j_mismatch_list.end());
             }
